+++ internal/provider/utils/constants.go
@@ -1,0 +1,9 @@
+// Copyright (c) Keeper Security, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package utils
+
+const (
+	EnterpriseManagedCompanySchemaAttributeDescription         = "Only applies to MSP accounts. Name or ID of the managed company to scope this resource or data source to. Omit to use the logged-in account context."
+	EnterpriseManagedCompanySchemaAttributeMarkdownDescription = "Only applies to **MSP accounts**. Name or ID of the managed company to scope this resource or data source to. Omit to use the logged-in account context."
+)